+++ app/layout.jsx
@@ -1,5 +1,5 @@
 import { Inter } from "next/font/google";
-import Nav from "./components/Nav";
+import Nav from "../components/Nav";
 import "./globals.css";
 import Providers from "./providers";
